changeListDemo to DataStructureDemo, remove stackDemo and TreeDemo.
