+++ CH2/stevenli/HW1/pycode/DataStructureDemo.py
@@ -1,0 +1,42 @@
+from collections import deque
+
+# This is a py class for data structure demo
+
+
+class listObj:
+    def __init__(self):
+        self = []
+
+
+class dequeObj:
+    def __init__(self):
+        self = []
+
+class stackObj:
+    def __init__(self):
+        self = []
+# create a tree class
+class treeObj:
+    def __init__(self, key):
+        self.left=None
+        self.right = None
+        self.val=key
+
+
+
+
+# create a main function to call above method and class
+def main():
+    print("This is main function")
+
+
+
+
+
+
+if __name__=='__main__':
+    main()
+
+
+
+
